initial commit - added cjs support, defined initial exports, moved validateCLioptions into app folder
