+++ package.json
@@ -3,14 +3,24 @@
     "version": "0.2.1",
     "description": "Your command center for controlling computers and services with natural language - connect once, command everything",
     "type": "module",
-    "main": "dist/src/app/index.js",
+    "main": "dist/src/core/index.cjs",
+    "module": "dist/src/core/index.js",
+    "types": "dist/src/core/index.d.ts",
     "bin": {
         "saiki": "./dist/src/app/index.js"
+    },
+    "exports": {
+        ".": {
+            "types": "./dist/src/core/index.d.ts",
+            "import": "./dist/src/core/index.js",
+            "require": "./dist/src/core/index.cjs"
+        },
+        "./package.json": "./package.json"
     },
     "scripts": {
         "clean": "npx rimraf dist",
         "prebuild": "",
-        "build": "npm run clean && npx tsc && npm run copy-client",
+        "build": "npm run clean && tsup src/core/index.ts --format esm,cjs --dts --out-dir dist/src/core && tsup src/app/index.ts --format esm --out-dir dist/src/app && npm run copy-client",
         "prepare": "husky",
         "copy-client": "npx rimraf public && mkdir public && npx copyfiles -f \"src/app/web/client/*\" public",
         "start": "node dist/src/app/index.js",
@@ -82,6 +92,7 @@
         "prettier": "^3.5.3",
         "supertest": "^7.1.0",
         "ts-node": "^10.9.2",
+        "tsup": "^8.4.0",
         "vitest": "^3.1.3"
     },
     "lint-staged": {
